requirements.txt
requests==2.32.5
pandas==3.0.1
pyarrow==22.0.0
python-dotenv==1.2.1
psycopg2-binary==2.9.11
boto3==1.40.37
